Changed structure of like functionality

+++ app/controllers/api/likes_controller.rb
@@ -12,7 +12,7 @@ class Api::LikesController < ApplicationController
   def create
     @like = Like.new(post_id: params[:id], user_id: current_user.id)
     if @like.save
-      render json: params[:id]
+      render json: {id: params[:id], liked: true}
     else
       render json: @like.errors.full_messages.join(', ')
     end
@@ -21,10 +21,14 @@ class Api::LikesController < ApplicationController
   def destroy
     @like = current_user.likes.find_by(post_id: params[:id])
 
-    if @like.destroy
-      render json: params[:id]
+    if @like
+      if @like.destroy
+        render json: {id: params[:id], liked: false}
+      else
+        render json: @like.errors.full_messages.join(', ')
+      end
     else
-      render json: @like.errors.full_messages.join(', ')
+      render json: "like does not exist"
     end
   end
